Added math structs

+++ micromouse_robot/include/micromouse_robot/rclcpp_range_sensor.hpp
@@ -4,14 +4,13 @@
 #include <sensor_msgs/msg/range.hpp>
 #include <string>
 #include <thread>
-
 #include "range_sensor_interface.hpp"
 
 namespace micromouse {
 class RclcppRangeSensor : public RangeSensorInterface, public rclcpp::Node {
    public:
     RclcppRangeSensor(const std::string name);
-    virtual double get_range() override final;
+    virtual float get_range() override final;
 
    private:
     rclcpp::Subscription<sensor_msgs::msg::Range>::SharedPtr range_sub_;
